tests/wpt/web-platform-tests/service-workers/service-worker/controller-on-disconnect.https.html

<!DOCTYPE html>
<title>Service Worker: Controller on load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
var t = async_test('controller is cleared on disconnected window');
t.step(function() {
    var url = 'resources/empty-worker.js';
    var scope = 'resources/blank.html';
    var registration;
    var controller;
    var frame;
    service_worker_unregister_and_register(t, url, scope)
      .then(t.step_func(function(swr) {
          registration = swr;
          return wait_for_state(t, registration.installing, 'activated');
        }))
      .then(t.step_func(function() {
          return with_iframe(scope)
        }))
      .then(t.step_func(function(f) {
          frame = f;
          var w = frame.contentWindow;
          var swc = w.navigator.serviceWorker;
          assert_true(swc.controller instanceof w.ServiceWorker,
                      'controller should be a ServiceWorker object');

          frame.remove();

          assert_equals(swc.controller, null,
                        'disconnected frame should not be controlled');

          service_worker_unregister_and_done(t, scope);
        }))
      .catch(unreached_rejection(t));
  });
</script>
</body>
