LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controller-on-disconnect.https-expected.txt


FAIL controller is cleared on disconnected window assert_equals: disconnected frame should not be controlled expected null but got object "[object ServiceWorker]"
